requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
# psycopg2==2.9.10  # vercel complains!
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Flask-Migrate==4.1.0
email_validator==1.3.1
email-normalize==0.2.1
Flask-JWT-Extended==4.7.1
flask-smorest==0.46.2
marshmallow-sqlalchemy==1.4.2
sqlakeyset==2.0.1746777265
